+++ server/routes/bothAdminStaff.js
@@ -129,112 +129,4 @@ router.patch('/account/:accountNumber', jwtStaffVerify, (req, res) => {
 	}
 });
 
-router.post('/transactions/:accountNumber/credit', jwtStaffVerify, (req, res) => {
-	let getUser = helper.togetUser(req);
-	if (getUser) {
-		//the field with "creditAmount" name
-		const getAmountcredit = req.body['creditAmount'] || 0.0;
-		//verify that the account number exist
-		const getacc = accounts.find(acc => acc.accountNumber === Number(req.params.accountNumber));
-		if (!getacc){
-			res.json({
-				"status":2004,
-				"error": `Cannot find a matching account number ${req.params.accountNumber}`
-			})
-		}
-		let initialAmount = parseFloat(getacc.balance);
-		const newAmount = parseFloat(getAmountcredit) + initialAmount;
-		//Replace the old balance with the new one
-		getacc.balance = newAmount;
-		// create a new transaction 
-		const newTransaction = {
-			"id": transactions.length + 1,
-	        "createdOn": new Date().toISOString(),
-			"transactionType": "credit",
-	        "accountNumber": getacc.accountNumber,
-			"cashier": getUser.id,
-			"amount": getAmountcredit,
-			"oldBalance": initialAmount,
-			"newBalance": getacc.balance,
-		}
-		// push the new transaction to the data
-		transactions.push(newTransaction);
-
-		// console.log(transactions); 
-		res.json({
-			"status": 1000,
-			"data": {
-				"transactionId": newTransaction.id,
-				"accountNumber": newTransaction.accountNumber,
-				"amount": newTransaction.amount,
-				"cashier": newTransaction.cashier,
-				"transactionType": newTransaction.transactionType,
-				"accountBalance": newTransaction.newBalance,
-				"oldBalance": newTransaction.oldBalance
-			} 
-		});
-
-	} else {
-		res.json({
-			"status": 1004,
-			"error": "Invalid User Stay Out!"
-		})
-	}
-});
-
-
-router.post('/transactions/:accountNumber/debit', jwtStaffVerify, (req, res) => {
-	let getUser = helper.togetUser(req);
-	if (getUser) {
-		//the field with "creditAmount" name
-		const getAmountcredit = req.body['debitAmount'] || 0.0;
-		//verify that the account number exist
-		const getacc = accounts.find(acc => acc.accountNumber === Number(req.params.accountNumber));
-		if (!getacc){
-			res.json({
-				"status":2004,
-				"error": `Cannot find a matching account number ${req.params.accountNumber}`
-			})
-		}
-		let initialAmount = parseFloat(getacc.balance);
-		const newAmount = initialAmount - parseFloat(getAmountcredit);
-		//Replace the old balance with the new one
-		getacc.balance = newAmount;
-		// create a new transaction 
-		const newTransaction = {
-			"id": transactions.length + 1,
-	        "createdOn": new Date().toISOString(),
-			"transactionType": "credit",
-	        "accountNumber": getacc.accountNumber,
-			"cashier": getUser.id,
-			"amount": getAmountcredit,
-			"oldBalance": initialAmount,
-			"newBalance": getacc.balance,
-		}
-		// push the new transaction to the data
-		transactions.push(newTransaction);
-
-		// console.log(transactions); 
-		res.json({
-			"status": 1000,
-			"data": {
-				"transactionId": newTransaction.id,
-				"accountNumber": newTransaction.accountNumber,
-				"amount": newTransaction.amount,
-				"cashier": newTransaction.cashier,
-				"transactionType": newTransaction.transactionType,
-				"accountBalance": newTransaction.newBalance,
-				"oldBalance": newTransaction.oldBalance
-			} 
-		});
-
-	} else {
-		res.json({
-			"status": 1004,
-			"error": "Invalid User Stay Out!"
-		})
-	}
-});
-
-
 module.exports = router;
